Change int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {99,30,5,100,78};
+        long[] sales = {99,30,5,100,78};
             SalesManager salesManager = new SalesManager(sales);
         System.out.println(salesManager.max());
     }
